+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [4.3.1](https://github.com/purple-technology/phoenix-components/compare/v4.3.0...v4.3.1) (2021-07-15)
+
+
+### Bug Fixes
+
+* selectpicker component - loop fix and checkbox styling fix ([69f18e7](https://github.com/purple-technology/phoenix-components/commit/69f18e731d9917d164141532f40246407f687871))
+
+## [4.3.0](https://github.com/purple-technology/phoenix-components/compare/v4.2.0...v4.3.0) (2021-07-14)
+
+
+### Features
+
+* datetime component - setTouched prop added ([729762e](https://github.com/purple-technology/phoenix-components/commit/729762e6e83d2b57271bca7124a098e58c39ed8d))
+
 ## [4.2.0](https://github.com/purple-technology/phoenix-components/compare/v4.1.0...v4.2.0) (2021-07-08)